TKinter/requirements.txt
mysql-connector==2.2.9
Pillow==9.1.1
ttkthemes==3.2.2
